# Some optimizations (cont)