@QuarticCat 2022-09-30
Refactor parents' representation.

**Speed**: 187%

**Memory**: 73%

```text
>>>>> BENCH START
Benchmark 1: /home/qc/.cargo/target/release/difft sample_files/slow_before.rs sample_files/slow_after.rs
  Time (mean ± σ):     600.1 ms ±   3.8 ms    [User: 567.1 ms, System: 31.5 ms]
  Range (min … max):   594.9 ms … 606.2 ms    10 runs
 
>>>>> BENCH END
>>>>> BENCH START

 Performance counter stats for '/home/qc/.cargo/target/release/difft sample_files/slow_before.rs sample_files/slow_after.rs':

            582.59 msec task-clock:u              #    0.999 CPUs utilized          
                 0      context-switches:u        #    0.000 /sec                   
                 0      cpu-migrations:u          #    0.000 /sec                   
             1,344      page-faults:u             #    2.307 K/sec                  
     1,976,757,963      cycles:u                  #    3.393 GHz                    
        35,702,044      stalled-cycles-frontend:u #    1.81% frontend cycles idle   
       760,550,837      stalled-cycles-backend:u  #   38.47% backend cycles idle    
     2,545,504,360      instructions:u            #    1.29  insn per cycle         
                                                  #    0.30  stalled cycles per insn
       472,339,368      branches:u                #  810.760 M/sec                  
        13,293,176      branch-misses:u           #    2.81% of all branches        

       0.582992026 seconds time elapsed

       0.552362000 seconds user
       0.030102000 seconds sys


>>>>> BENCH END
~/.cargo/target/release/difft sample_files/slow_{before,after}.rs > /dev/null   0.50s  user 0.03s system 99% cpu 0.528 total
avg shared (code):         0 KB
avg unshared (data/stack): 0 KB
total (sum):               0 KB
max memory:                401 KB
page faults from disk:     0
other page faults:         1418
```