@QuarticCat 2022-09-30
Compress `EnteredDelimiter`.

**Speed**: 189%

**Memory**: 71%

```text
>>>>> BENCH START
Benchmark 1: /home/qc/.cargo/target/release/difft sample_files/slow_before.rs sample_files/slow_after.rs
  Time (mean ± σ):     591.4 ms ±  11.1 ms    [User: 559.8 ms, System: 30.9 ms]
  Range (min … max):   582.6 ms … 618.6 ms    10 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
>>>>> BENCH END
>>>>> BENCH START

 Performance counter stats for '/home/qc/.cargo/target/release/difft sample_files/slow_before.rs sample_files/slow_after.rs':

            576.92 msec task-clock:u              #    0.999 CPUs utilized          
                 0      context-switches:u        #    0.000 /sec                   
                 0      cpu-migrations:u          #    0.000 /sec                   
             1,339      page-faults:u             #    2.321 K/sec                  
     1,960,597,413      cycles:u                  #    3.398 GHz                    
        47,556,040      stalled-cycles-frontend:u #    2.43% frontend cycles idle   
       711,186,359      stalled-cycles-backend:u  #   36.27% backend cycles idle    
     2,530,607,392      instructions:u            #    1.29  insn per cycle         
                                                  #    0.28  stalled cycles per insn
       467,326,782      branches:u                #  810.038 M/sec                  
        13,318,875      branch-misses:u           #    2.85% of all branches        

       0.577286658 seconds time elapsed

       0.529902000 seconds user
       0.046657000 seconds sys


>>>>> BENCH END
~/.cargo/target/release/difft sample_files/slow_{before,after}.rs > /dev/null   0.51s  user 0.02s system 99% cpu 0.530 total
avg shared (code):         0 KB
avg unshared (data/stack): 0 KB
total (sum):               0 KB
max memory:                391 KB
page faults from disk:     0
other page faults:         1408
```