@QuarticCat 2022-09-30
Reserve vec capacity.

**Speed**: 202%

**Memory**: 71%

```text
>>>>> BENCH START
Benchmark 1: /home/qc/.cargo/target/release/difft sample_files/slow_before.rs sample_files/slow_after.rs
  Time (mean ± σ):     555.5 ms ±   5.2 ms    [User: 521.3 ms, System: 33.5 ms]
  Range (min … max):   549.2 ms … 564.5 ms    10 runs
 
>>>>> BENCH END
>>>>> BENCH START

 Performance counter stats for '/home/qc/.cargo/target/release/difft sample_files/slow_before.rs sample_files/slow_after.rs':

            560.13 msec task-clock:u              #    0.999 CPUs utilized          
                 0      context-switches:u        #    0.000 /sec                   
                 0      cpu-migrations:u          #    0.000 /sec                   
             1,339      page-faults:u             #    2.391 K/sec                  
     1,898,772,574      cycles:u                  #    3.390 GHz                    
        39,751,378      stalled-cycles-frontend:u #    2.09% frontend cycles idle   
       677,732,625      stalled-cycles-backend:u  #   35.69% backend cycles idle    
     2,385,056,239      instructions:u            #    1.26  insn per cycle         
                                                  #    0.28  stalled cycles per insn
       439,507,344      branches:u                #  784.657 M/sec                  
        13,799,727      branch-misses:u           #    3.14% of all branches        

       0.560540216 seconds time elapsed

       0.526526000 seconds user
       0.033323000 seconds sys


>>>>> BENCH END
~/.cargo/target/release/difft sample_files/slow_{before,after}.rs > /dev/null   0.48s  user 0.03s system 99% cpu 0.511 total
avg shared (code):         0 KB
avg unshared (data/stack): 0 KB
total (sum):               0 KB
max memory:                391 KB
page faults from disk:     0
other page faults:         1404
```