@QuarticCat 2022-09-30
Compress seen map.

Here `hashbrown` is introduced since it has a `get_key_value_mut` method.

**Speed**:  207%

**Memory**: 68%

```text
>>>>> BENCH START
Benchmark 1: /home/qc/.cargo/target/release/difft sample_files/slow_before.rs sample_files/slow_after.rs
  Time (mean ± σ):     540.0 ms ±  10.0 ms    [User: 509.0 ms, System: 30.0 ms]
  Range (min … max):   529.4 ms … 557.1 ms    10 runs
 
>>>>> BENCH END
>>>>> BENCH START

 Performance counter stats for '/home/qc/.cargo/target/release/difft sample_files/slow_before.rs sample_files/slow_after.rs':

            520.83 msec task-clock:u              #    0.999 CPUs utilized          
                 0      context-switches:u        #    0.000 /sec                   
                 0      cpu-migrations:u          #    0.000 /sec                   
             1,334      page-faults:u             #    2.561 K/sec                  
     1,761,757,697      cycles:u                  #    3.383 GHz                    
        35,457,864      stalled-cycles-frontend:u #    2.01% frontend cycles idle   
       601,687,452      stalled-cycles-backend:u  #   34.15% backend cycles idle    
     2,244,942,485      instructions:u            #    1.27  insn per cycle         
                                                  #    0.27  stalled cycles per insn
       418,526,690      branches:u                #  803.569 M/sec                  
        13,789,606      branch-misses:u           #    3.29% of all branches        

       0.521212130 seconds time elapsed

       0.483899000 seconds user
       0.036709000 seconds sys


>>>>> BENCH END
~/.cargo/target/release/difft sample_files/slow_{before,after}.rs > /dev/null   0.43s  user 0.04s system 99% cpu 0.469 total
avg shared (code):         0 KB
avg unshared (data/stack): 0 KB
total (sum):               0 KB
max memory:                375 KB
page faults from disk:     0
other page faults:         1396
```